refactor: move JsonMapConverter into the shared db package
The compute-service and storage-service each carried a byte-identical copy of this JPA JSON-map AttributeConverter; this keeps a single copy in org.apache.airavata.db alongside JsonListConverter and points the two @Convert references (ComputeJobSubmissionEntity, StorageDataMovementEntity) at it. Behavior-preserving and the reactor builds green.
