--- airavata-api/compute-service/src/main/java/org/apache/airavata/compute/model/JsonMapConverter.java
@@ -1,51 +1,0 @@
-/**
-*
-* Licensed to the Apache Software Foundation (ASF) under one
-* or more contributor license agreements. See the NOTICE file
-* distributed with this work for additional information
-* regarding copyright ownership. The ASF licenses this file
-* to you under the Apache License, Version 2.0 (the
-* "License"); you may not use this file except in compliance
-* with the License. You may obtain a copy of the License at
-*
-* http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied. See the License for the
-* specific language governing permissions and limitations
-* under the License.
-*/
-package org.apache.airavata.compute.model;
-
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.persistence.AttributeConverter;
-import jakarta.persistence.Converter;
-import java.util.Map;
-
-@Converter
-public class JsonMapConverter implements AttributeConverter<Map<String, Object>, String> {
-    private static final ObjectMapper MAPPER = new ObjectMapper();
-
-    @Override
-    public String convertToDatabaseColumn(Map<String, Object> attribute) {
-        if (attribute == null) return null;
-        try {
-            return MAPPER.writeValueAsString(attribute);
-        } catch (Exception e) {
-            throw new IllegalArgumentException("JSON serialization failed", e);
-        }
-    }
-
-    @Override
-    public Map<String, Object> convertToEntityAttribute(String dbData) {
-        if (dbData == null) return null;
-        try {
-            return MAPPER.readValue(dbData, new TypeReference<>() {});
-        } catch (Exception e) {
-            throw new IllegalArgumentException("JSON deserialization failed", e);
-        }
-    }
-}
