+++ airavata-api/storage-service/src/main/java/org/apache/airavata/storage/model/StorageDataMovementEntity.java
@@ -23,6 +23,7 @@ import jakarta.persistence.*;
 import java.io.Serializable;
 import java.sql.Timestamp;
 import java.util.Map;
+import org.apache.airavata.db.JsonMapConverter;
 import org.apache.airavata.model.data.movement.proto.SecurityProtocol;
 
 /**
